requirements.txt
Pillow
Click
Crayons

# Testing
black
codecov
mypy
pytest
pytest-cov